Supprimer un gabs

+++ api/src/main/java/com/supinfo/gabbler/server/repository/GabsRepository.java
@@ -4,5 +4,5 @@ import com.supinfo.gabbler.server.entity.Gabs;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface GabsRepository extends JpaRepository<Gabs, Integer>, JpaSpecificationExecutor {
+public interface GabsRepository extends JpaRepository<Gabs, Long>, JpaSpecificationExecutor {
 }
